# oapi/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Article Analyzer
  description: Text Analysis Application
  version: 0.1.0
servers:
  - url: http://localhost:8000/
paths:
...
  /article:
    get:
      summary: Request to get article content
      operationId: article
      parameters:
        - name: url
          in: query
          description: URL of the article
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          example: "Error message"
      required:
        - message
    ExtractResponse:
      type: object
      properties:
        content:
          type: string
          example: "Extracted content here"
      required:
        - content
...